fix: change usd type to integer
According to the webhook spec, it should be an integer, not a float.

+++ api/types.rs
@@ -20,5 +20,5 @@ pub struct Sponsor {
 
 #[derive(Debug, Deserialize)]
 pub struct Tier {
-    pub monthly_price_in_dollars: f64,
+    pub monthly_price_in_dollars: i64,
 }
